Migration: mention the new migration-reference/0
Back in "f3786c9 Migration: mention the new migration-reference/0" we
added an info about what migration-reference runs are and how to use
them. This was, by accident, removed in: "ac4f624 ProposedMigration:
Explain excuses and suggest course of actions".

Let us add that back, in a more elaborate way, as it has become a quite
common task for anyone dealing with proposed migration nowadays.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>