Add rel-notes check to MRE preparation
We really spend a lot of effort in preparing and processing SRU
exceptions, but it is only a little investment to be a bit extra
careful on those updates.

Let us add a section mentioning how MREs are prepared in general
and in there a hint about the benefit of not only wrapping the code
into an update but conciously checking the release notes / changelog
in regard to their match of the SRU philosophy.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>