Fix missleading use of `groupadd`
This commit replaces `groupadd` with `adduser`.

The current description implies that `groupadd` will add the user to the
specified groups if the user notices that the output of `groups my_user`
does not contain `lxd`, `sbuild` or `libvirt`.
Instead `groupadd` will create these groups, but does not add the
user to these groups as described.