CommittingChanges.md, VersionStrings.md: Clarify upload versioning
We have had too many cases where referring to the security wiki page
that served as the gold standard was not sufficient. Especially around
going ahead via upstream merges and even more so backports and pushing
new versions as MREs it needs to be spelled out to avoid common mistakes
from re-occuring.

Review feedback was awesome and helped to make this even more complete
and useful than I thought. Big thanks to all reviewers and co-authors
that eventually worked on this with me.

This splits the short section we had into a page of its own which
outlines all related cases and therefore contains:
- Introduction to the problem
- Uploading changes to Ubuntu
- Merges from Debian
- Merging from upstream going ahead of Debian.
- Stable release updates
- Native packages
- No change rebuilds
- Backports of upstream versions (MRE).
- Backports from -devel (same on all releases).
- Undoing mistakes (+really)

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Co-authored-by: James Falcon <therealfalcon@gmail.com>
Co-authored-by: Grant Orndorff <orndorffgrant@protonmail.com>
Co-authored-by: Sally <sally.makin@canonical.com>
Co-authored-by: Andreas Hasenack <andreas@canonical.com>