Avoid referring to a stale local branch
ubuntu/devel might be a local branch and that might have gotten out of date. People could waste time until they spot to work on the wrong version.

Instead refer to the pkg remote which is the latest state of the importer and should always be a better choice for merge start/finish.