Setup.md: clean common tools list
It is ok that we have some rearely used tools like debmake, or
those more likely to be present in an actual build schroot or container
like pkg-config, but that is fine.
The only one that really is outdated, no more needed nowadays and we
should not recomment it is net-tools.

Not so helpful is to list duplicates that are pulled in via dependencies
anyway. ubuntu-dev-tools already brings in devscripts and dpkg-dev.
libvirt-daemon-system brings in libvirt-daemon, ovmf, and qemu-system.
sbuild-launchpad-chroot brings sbuild.

Make the list more concise by only listing the top level entries that
are needed.